Split up FieldFactory and Mapper
